refactor: LottoGame 도메인 정의
