+++ src/main/kotlin/model/lotto/LottoTicket.kt
@@ -1,41 +1,14 @@
 package model.lotto
 
-enum class TicketType {
-    Manual,
-    Auto,
-}
-
-data class LottoTicket(
-    val numbers: LottoNumbers,
-    val type: TicketType,
-) {
+// Lotto Game 을 감싸는 일급 컬렉션
+class LottoTicket(private val games: List<LottoGame>) {
     companion object {
-        fun of(
-            numbers: LottoNumbers,
-            type: TicketType,
-        ): LottoTicket {
-            return LottoTicket(numbers, type)
+        fun of(games: List<LottoGame>): LottoTicket {
+            return LottoTicket(games)
         }
     }
 
     override fun toString(): String {
-        return numbers.toString()
-    }
-}
-
-object LottoTicketGenerator {
-    fun generate(count: Int): List<LottoTicket> {
-        val allPossibleNumbers = (1..45).toList()
-        return List(count) {
-            val selectedNumbers = allPossibleNumbers.shuffled().take(6).sorted()
-            val lottoNumbers = LottoNumbers.from(selectedNumbers)
-            LottoTicket.of(lottoNumbers, TicketType.Auto)
-        }
-    }
-
-    fun generate(manualNumbers: List<LottoNumbers>): List<LottoTicket> {
-        return manualNumbers.map { lottoNumbers ->
-            LottoTicket.of(lottoNumbers, TicketType.Manual)
-        }
+        return games.joinToString("\n")
     }
 }
